lib/stdio: Handle 64bits off_t correctly
